Add Line Graph
