+++ App.tsx
@@ -4,7 +4,7 @@ import Constants from "expo-constants";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import Home from "./container/home/Home";
 import React from "react";
-import Graph from "./container/home/Graph";
+import Graph from "./container/home/GraphSection";
 import { NavigationContainer } from "@react-navigation/native";
 import { Ionicons } from "@expo/vector-icons";
 
@@ -40,26 +40,7 @@ export default function App() {
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1,
     flexDirection: "column",
     alignItems: "center",
-  },
-
-  headerSection: {
-    flex: 1,
-    flexDirection: "row",
-    marginTop: Constants.statusBarHeight,
-    backgroundColor: "#1A434D",
-  },
-
-  graphSection: {
-    flex: 2,
-    flexDirection: "row",
-    backgroundColor: "white",
-  },
-
-  tasks: {
-    flex: 2,
-    backgroundColor: "white",
   },
 });
