+++ package.json
@@ -22,8 +22,10 @@
     "react-devtools-core": "4.13.0",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
+    "react-native-chart-kit": "^6.12.0",
     "react-native-safe-area-context": "4.2.4",
     "react-native-safe-area-view": "^1.1.1",
+    "react-native-svg": "^12.4.0",
     "react-native-vector-icons": "^9.2.0",
     "react-native-web": "0.17.7"
   },
